+++ i3_rofi/cli.py
@@ -3,12 +3,11 @@ from i3_rofi import I3Rofi
 
 
 def run():
-    i3rofi = I3Rofi()
     parser = argparse.ArgumentParser(
         description='Provides rofi menus to interact with i3')
     parser.add_argument(
         "-m", "--menu", dest="menu", required=True,
-        choices=i3rofi.menus,
+        choices=I3Rofi.menus,
         help="Menu to display"
     )
     parser.add_argument(
@@ -16,4 +15,5 @@ def run():
         help="Display debug infos"
     )
     args = parser.parse_args()
-    i3rofi.run(args)
+    i3rofi = I3Rofi(menu=args.menu, debug=args.debug)
+    i3rofi()