+++ i3_rofi/i3_rofi.py
@@ -1,22 +1,59 @@
 # -*- coding: utf-8 -*-
 import sys
 import errno
-import i3
-import subprocess
 from utils import which
 from utils import safe_list_get
 from utils import rofi
+from utils import i3_command
+from utils import select
+from utils import select_output
+from utils import select_workspace
+from utils import select_window
+from utils import select_bar
 import i18n
-
-GO_BACK_OPTION = '<- go back'
-GO_BACK_SIGNAL = '<go back>'
-ROFI_PREFIX = '(i3-rofi)'
 
 _ = i18n.language.gettext
 DEFAULT_TITLE = _('Select:')
 
+COMMANDS_ACTIONS = {
+    'sticky': [
+        'enable', 'disable', 'toggle'],
+    'fullscreen': [
+        'enable', 'disable', 'toggle'],
+    'floating': [
+        'enable', 'disable', 'toggle'],
+    'bar_hidden_state': [
+        'hide', 'show', 'toggle'],
+    'bar_mode': [
+        'dock', 'hide', 'invisible', 'toggle'],
+    'layout': [
+        'default', 'tabbed', 'stacking', 'splitv', 'splith']
+}
+
 
 class I3Rofi(object):
+    menus = [
+        'go_to_workspace',
+        'move_window_to_workspace',
+        'move_window_to_this_workspace',
+        'move_workspace_to_output',
+        'rename_workspace',
+        'window_actions',
+        'workspace_actions',
+        'scratchpad_actions',
+        'bar_actions',
+        'layout',
+    ]
+
+    def __init__(self, menu, debug=False):
+        if menu not in self.menus:
+            sys.exit(errno.EINVAL)
+        self.menu_fnc = getattr(self, menu)
+        self.debug = debug
+
+    def __call__(self):
+        self.check_rofi()
+        self.menu_fnc()
 
     def check_rofi(self):
         '''Check if rofi is available.'''
@@ -25,141 +62,136 @@ class I3Rofi(object):
         else:
             sys.exit(errno.EINVAL)
 
-    def _i3_get_workspaces(self):
-        return i3.get_workspaces()
-
-    def _i3_get_outputs(self):
-        return i3.get_outputs()
-
-    def _i3_get_windows(self):
-        return i3.filter(**{'nodes': [], 'type': 'con'})
-
-    def _i3_get_focused_workspace(self):
-        return i3.filter(tree=i3.get_workspaces(), focused=True)[0]
-
-    def _i3_get_unfocused_workspaces(self):
-        return i3.filter(tree=i3.get_workspaces(), focused=False)
-
-    def _i3_get_active_outputs(self):
-        return i3.filter(tree=i3.get_outputs(), active=True)
-
-    def _i3_get_focused_output(self):
-        ws = self._i3_get_focused_workspace()
-        return ws['output']
-
-    def _i3_get_unfocused_outputs(self):
-        active_outputs = self._i3_get_active_outputs()
-        focused_output = self._i3_get_focused_output()
-        active_outputs.pop(active_outputs.index(focused_output))
-        return active_outputs
-
-    def _select_workspace(self, title=DEFAULT_TITLE):
-        entries_list = self._i3_get_workspaces()
-        entries_list = sorted(entries_list, key=lambda x: x.get('name'))
-        labels = []
-        for ws in entries_list:
-            current = ws['focused'] and ' (current)' or ''
-            label = '{name} {current}'.format(
-                name=ws['name'], current=current)
-            labels.append(label)
-        idx = rofi(labels, title=title)
-        return safe_list_get(entries_list, idx, None)
-
-    def _select_output(self, title=DEFAULT_TITLE,
-                       unfocused_only=True):
-        entries_list = self._i3_get_active_outputs()
-        entries_list = sorted(entries_list, key=lambda x: x.get('name'))
-        labels = []
-        focused_output = self._i3_get_focused_output()
-        for i, out in enumerate(entries_list):
-            current = out['name'] == focused_output \
-                and '(current)' or ''
-            label = '{idx}: {name} {current}'.format(
-                idx=i, name=out['name'], current=current)
-            labels.append(label)
-        idx = rofi(labels, title=title)
-        return safe_list_get(entries_list, idx, None)
-
-    def _select_window(self, title=DEFAULT_TITLE):
-        entries = []
-        entries_list = self._i3_get_windows()
-        for win in entries_list:
-            entry = '{winclass}\t{title}'.format(
-                winclass=win['window_properties']['class'],
-                title=win['name'].encode('utf-8'))
-            entries.append(entry)
-        ordered_list = [
-            '{idx}: {entry}'.format(idx=i + 1, entry=e)
-            for i, e in enumerate(entries)]
-        idx = rofi(
-            ordered_list,
-            title=title)
-        return safe_list_get(entries_list, idx, None)
-
     def go_to_workspace(self, debug=False):
-        ws = self._select_workspace(_('Go to workspace:'))
-        if not ws:
-            return GO_BACK_SIGNAL
-        if debug:
-            print 'i3-msg workspaces "{ws}"'.format(ws=ws.get('name'))
-        return i3.workspace(ws.get('name'))
+        ws = select_workspace(_('Go to workspace:'))
+        cmd = 'workspace "{ws}"'.format(ws=ws.name)
+        return i3_command(cmd, debug=debug)
 
     def move_window_to_workspace(self, debug=False):
-        ws = self._select_workspace(_('Move window to workspace:'))
-        if not ws:
-            return GO_BACK_SIGNAL
-        CMD = 'window to workspace "{ws}"'.format(ws=ws.get('name'))
-        if debug:
-            print CMD
-        return i3.move(CMD)
+        ws = select_workspace(_('Move window to workspace:'))
+        cmd = 'move window to workspace "{ws}"'.format(ws=ws.name)
+        return i3_command(cmd, debug=debug)
 
     def move_workspace_to_output(self, debug=False):
-        out = self._select_output(_('Move active workspace to output:'))
-        if not out:
-            return GO_BACK_SIGNAL
-        CMD = 'workspace to output "{output}"'.format(output=out.get('name'))
-        if debug:
-            print CMD
-        return i3.move(CMD)
+        out = select_output(_('Move active workspace to output:'))
+        cmd = 'move workspace to output "{output}"'.format(output=out.name)
+        return i3_command(cmd, debug=debug)
 
     def rename_workspace(self, debug=False):
-        ws = self._i3_get_focused_workspace()
+        ws = select_workspace(filter_fnc=lambda x: x.focused)
         choice = rofi(
-            [ws['name']], _('Rename workspace:'), **{'format': 's'})
-        if not choice or choice == GO_BACK_OPTION:
-            return GO_BACK_SIGNAL
-        CMD = 'workspace to "{output}"'.format(output=choice)
-        if debug:
-            print CMD
-        return i3.rename(CMD)
+            [ws.name.encode('utf-8')], _('Rename workspace:'), **{'format': 's'})
+        if not choice:
+            return
+        cmd = 'rename workspace to "{output}"'.format(output=choice)
+        return i3_command(cmd, debug=debug)
 
     def move_window_to_this_workspace(self, debug=False):
-        win = self._select_window(
+        win = select_window(
             _('Select a window to move to this workspace:'))
-        if not win:
-            return GO_BACK_SIGNAL
-        CMD = '[class="{winclass}"] move window to workspace current'.format(
-            winclass=win['window_properties']['class'])
-        if debug:
-            print CMD
-        return i3.command(CMD)
+        cmd = '[id="{id}"] move window to workspace current'.format(
+            id=win.window)
+        return i3_command(cmd, debug=debug)
 
-    def window_actions(self, debug=False):
+    def floating(self, action=None, debug=False):
+        """ http://i3wm.org/docs/userguide.html#_manipulating_layout
+        """
+        actions = COMMANDS_ACTIONS['floating']
+        if not action or action not in actions:
+            action = select(actions, title='action:')
+        cmd = 'floating {action}'.format(action=action)
+        return i3_command(cmd, debug=debug)
+
+    def fullscreen(self, action=None, debug=False):
+        """ http://i3wm.org/docs/userguide.html#_manipulating_layout
+        """
+        actions = COMMANDS_ACTIONS['fullscreen']
+        if not action or action not in actions:
+            action = select(actions, title='action:')
+        cmd = 'fullscreen {action}'.format(action=action)
+        return i3_command(cmd, debug=debug)
+
+    def sticky(self, action=None, debug=False):
+        """ http://i3wm.org/docs/userguide.html#_sticky_floating_windows
+        """
+        actions = COMMANDS_ACTIONS['sticky']
+        if not action or action not in actions:
+            action = select(actions, title='action:')
+        cmd = 'sticky {action}'.format(action=action)
+        return i3_command(cmd, debug=debug)
+
+    def kill(self, debug=False):
+        cmd = 'kill'
+        return i3_command(cmd, debug=debug)
+
+    def move_to_scratchpad(self, debug=False):
+        cmd = 'move to scratchpad'
+        return i3_command(cmd, debug=debug)
+
+    def scratchpad_show(self, debug=False):
+        win = select_window(
+            _('Select a window to show:'),
+            scratchpad=True)
+        cmd = '[id="{id}"] scratchpad show'.format(id=win.window)
+        return i3_command(cmd, debug=debug)
+
+    def bar_hidden_state(self, action=None, debug=False):
+        """ http://i3wm.org/docs/userguide.html#_i3bar_control
+        """
+        bar_id = select_bar(_('Select bar:'))
+        actions = COMMANDS_ACTIONS['bar_hidden_state']
+        if not action or action not in actions:
+            action = select(actions, title='action:')
+        cmd = 'bar hidden_state {action} "{bar_id}"'.format(
+            action=action, bar_id=bar_id)
+        return i3_command(cmd, debug=debug)
+
+    def bar_mode(self, action=None, debug=False):
+        """ http://i3wm.org/docs/userguide.html#_i3bar_control
+        """
+        bar_id = select_bar(_('Select bar:'))
+        actions = COMMANDS_ACTIONS['bar_mode']
+        if not action or action not in actions:
+            action = select(actions, title='action:')
+        cmd = 'bar mode {action} "{bar_id}"'.format(
+            action=action, bar_id=bar_id)
+        return i3_command(cmd, debug=debug)
+
+    def layout(self, action=None, debug=False):
+        """ http://i3wm.org/docs/userguide.html#_manipulating_layout
+        """
+        actions = COMMANDS_ACTIONS['layout']
+        if not action or action not in actions:
+            action = select(actions, title='action:')
+        cmd = 'layout {action}'.format(action=action)
+        return i3_command(cmd, debug=debug)
+
+    def bar_actions(self, debug=False):
+        """ http://i3wm.org/docs/userguide.html#_i3bar_control
+        """
         actions = [
-            {'title': _('Move window to workspace:'),
-             'callback': self.move_window_to_workspace},
-            {'title': _('Floating (toggle)'),
-             'callback': lambda x: i3.floating('toggle')},
-            {'title': _('Fullscreen (toggle)'),
-             'callback': lambda x: i3.fullscreen()},
-            {'title': _('Sticky (toggle)'),
-             'callback': lambda x: i3.sticky('toggle')},
-            {'title': _('Move to Scratchpad'),
-             'callback': lambda x: i3.move('scratchpad')},
-            {'title': _('Move window to this workspace'),
-             'callback': self.move_window_to_this_workspace},
-            {'title': _('Quit'),
-             'callback': lambda x: i3.kill()}
+            {'title': _('hidden_state'),
+             'callback': self.bar_hidden_state},
+            {'title': _('mode'),
+             'callback': self.bar_mode},
+        ]
+        entries = [
+            '%s: %s' % (idx + 1, i['title'])
+            for idx,i in enumerate(actions)]
+        idx = rofi(entries, _('Bar actions:'))
+        action = safe_list_get(actions, idx, None)
+        callback = action['callback']
+        callback(debug=debug)
+        sys.exit()
+
+    def scratchpad_actions(self, debug=False):
+        """ http://i3wm.org/docs/userguide.html#_scratchpad
+        """
+        actions = [
+            {'title': _('Move current window to scratchpad'),
+             'callback': self.move_to_scratchpad},
+            {'title': _('Show window from scratchpad'),
+             'callback': self.scratchpad_show},
         ]
         entries = [
             '%s: %s' % (idx + 1, i['title'])
@@ -167,12 +199,36 @@ class I3Rofi(object):
         idx = rofi(entries, _('Window actions:'))
         action = safe_list_get(actions, idx, None)
         callback = action['callback']
-        res = callback(debug=debug)
-        if res == GO_BACK_SIGNAL:
-            self.window_actions()
-        return
+        callback(debug=debug)
+        sys.exit()
 
-    def workspace_actions(self, debug=True):
+    def window_actions(self, debug=False):
+        actions = [
+            {'title': _('Move window to workspace:'),
+             'callback': self.move_window_to_workspace},
+            {'title': _('Floating (toggle)'),
+             'callback': self.floating},
+            {'title': _('Fullscreen (toggle)'),
+             'callback': self.fullscreen},
+            {'title': _('Sticky'),
+             'callback': self.sticky},
+            {'title': _('Move to Scratchpad'),
+             'callback': self.move_to_scratchpad},
+            {'title': _('Move window to this workspace'),
+             'callback': self.move_window_to_this_workspace},
+            {'title': _('Quit'),
+             'callback': self.kill}
+        ]
+        entries = [
+            '%s: %s' % (idx + 1, i['title'])
+            for idx,i in enumerate(actions)]
+        idx = rofi(entries, _('Window actions:'))
+        action = safe_list_get(actions, idx, None)
+        callback = action['callback']
+        callback(debug=debug)
+        sys.exit()
+
+    def workspace_actions(self, debug=False):
         actions = [
             {'title': _('Go to workspace:'),
              'callback': self.go_to_workspace},
@@ -189,27 +245,5 @@ class I3Rofi(object):
         idx = rofi(entries, _('Workspace actions:'))
         action = safe_list_get(actions, idx, None)
         callback = action['callback']
-        res = callback()
-        if res == GO_BACK_SIGNAL:
-            self.workspace_actions()
-        return
-
-    @property
-    def menus(self):
-        menus = [
-            'go_to_workspace',
-            'move_window_to_workspace',
-            'move_window_to_this_workspace',
-            'move_workspace_to_output',
-            'rename_workspace',
-            'window_actions',
-            'workspace_actions',
-        ]
-        return menus
-
-    def run(self, cli_args):
-        self.check_rofi()
-        if cli_args.menu not in self.menus:
-            sys.exit(errno.EINVAL)
-        menu_fnc = getattr(self, cli_args.menu)
-        menu_fnc(debug=cli_args.debug)
+        callback(debug=debug)
+        sys.exit()