+++ docs/HISTORY.rst
@@ -1,7 +1,21 @@
 Changelog
 =========
 
-1.0dev (unreleased)
--------------------
+1.1 (unreleased)
+----------------
+
+- major code styles
+- add scratchpad actions
+- add bar actions
+- add layout action
+- switch from i3-py to i3ipc as base since it's more powerful and updated
+  [giacomos]
+
+1.0 (2016-02-18)
+----------------
 
 - Initial release
+- included menus are: go_to_workspace, move_window_to_workspace,
+  move_window_to_this_workspace, move_workspace_to_output, rename_workspace,
+  window_actions, workspace_actions
+  [giacomos]